Fix logging init.
