+++ adapters/ds/gst_plugins/python/always_on_rtsp_frame_processor.py
@@ -9,7 +9,8 @@ from adapters.ds.sinks.always_on_rtsp.last_frame import LastFrame
 from adapters.ds.sinks.always_on_rtsp.timestamp_overlay import TimestampOverlay
 from savant.deepstream.opencv_utils import nvds_to_gpu_mat
 from savant.gstreamer import Gst, GstBase, GObject
-from savant.gstreamer.utils import LoggerMixin, propagate_gst_setting_error
+from savant.gstreamer.utils import propagate_gst_setting_error
+from savant.utils.logging import LoggerMixin
 
 
 class Mode(Enum):
