+++ adapters/ds/gst_plugins/python/always_on_rtsp_frame_sink.py
@@ -6,7 +6,8 @@ from adapters.ds.sinks.always_on_rtsp.last_frame import LastFrame
 from adapters.ds.sinks.always_on_rtsp.timestamp_overlay import TimestampOverlay
 from savant.deepstream.opencv_utils import nvds_to_gpu_mat
 from savant.gstreamer import Gst, GstBase, GObject
-from savant.gstreamer.utils import LoggerMixin, propagate_gst_setting_error
+from savant.gstreamer.utils import propagate_gst_setting_error
+from savant.utils.logging import LoggerMixin
 
 CAPS = Gst.Caps.from_string('video/x-raw(memory:NVMM), format=RGBA')
