+++ adapters/gst/gst_plugins/python/fps_meter.py
@@ -2,7 +2,7 @@ from enum import Enum
 from typing import Any, Optional
 
 from savant.gstreamer import GObject, Gst, GstBase
-from savant.gstreamer.utils import LoggerMixin
+from savant.utils.logging import LoggerMixin
 from savant.utils.fps_meter import FPSMeter
