+++ adapters/gst/gst_plugins/python/zeromq_sink.py
@@ -9,7 +9,8 @@ from gst_plugins.python.zeromq_properties import (
     ZEROMQ_PROPERTIES,
 )
 from savant.gstreamer import GObject, Gst, GstBase
-from savant.gstreamer.utils import LoggerMixin, propagate_gst_setting_error
+from savant.gstreamer.utils import propagate_gst_setting_error
+from savant.utils.logging import LoggerMixin
 from savant.utils.zeromq import (
     Defaults,
     SenderSocketTypes,
