+++ gst_plugins/python/avro_video_decode_bin.py
@@ -9,7 +9,8 @@ from dataclasses import dataclass
 from gst_plugins.python.avro_video_demux import AVRO_VIDEO_DEMUX_PROPERTIES
 from savant.gstreamer import GLib, GObject, Gst  # noqa:F401
 from savant.gstreamer.codecs import Codec, CODEC_BY_CAPS_NAME
-from savant.gstreamer.utils import LoggerMixin, on_pad_event, pad_to_source_id
+from savant.gstreamer.utils import on_pad_event, pad_to_source_id
+from savant.utils.logging import LoggerMixin
 from savant.utils.platform import is_aarch64
 
 OUT_CAPS = Gst.Caps.from_string('video/x-raw(memory:NVMM);video/x-raw')
