+++ gst_plugins/python/zeromq_source_bin.py
@@ -5,7 +5,7 @@ from gst_plugins.python.avro_video_decode_bin import (
 )
 from gst_plugins.python.zeromq_src import ZEROMQ_SRC_PROPERTIES
 from savant.gstreamer import GObject, Gst
-from savant.gstreamer.utils import LoggerMixin
+from savant.utils.logging import LoggerMixin
 
 
 class ZeroMQSourceBin(LoggerMixin, Gst.Bin):
