+++ savant/entrypoint/main.py
@@ -1,10 +1,9 @@
 """Module entrypoint function."""
-import logging
 from savant.config import ModuleConfig
 from savant.gstreamer import Gst
 from savant.gstreamer.runner import GstPipelineRunner
 from savant.deepstream.pipeline import NvDsPipeline
-from savant.utils.logging import init_logging, update_logging
+from savant.utils.logging import init_logging, update_logging, get_logger
 from savant.utils.sink_factories import sink_factory
 
 
@@ -23,7 +22,7 @@ def main(config_file_path: str):
 
     # reconfigure savant logger with updated loglevel
     update_logging(config.parameters['log_level'])
-    logger = logging.getLogger('savant')
+    logger = get_logger('savant')
 
     # possible exceptions will cause app to crash and log error by default
     # no need to handle exceptions here
